requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiogram>=3.4.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.0
greenlet>=3.0.0
psycopg2-binary>=2.9.9
